backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai>=1.50.0
anthropic>=0.40.0
redis==5.0.1
nest-asyncio==1.5.8
python-multipart==0.0.6
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
httpx==0.25.1